Unifies mobile base robot
