+++ examples/mobile_base.py
@@ -5,12 +5,12 @@ import numpy as np
 
 def main():
     env = gym.make('mobile-base-acc-v0', render=True, dt=0.01)
-    defaultAction = [0.1]
+    defaultAction = [0.4]
     n_episodes = 1
     n_steps = 1000
     cumReward = 0.0
     for e in range(n_episodes):
-        ob = env.reset()
+        ob = env.reset(pos=np.array([-2.0]), vel=np.array([2.0]))
         print("Starting episode")
         for i in range(n_steps):
             action = env.action_space.sample()
