+++ mobileBase/envs/acc.py
@@ -1,119 +1,19 @@
 import numpy as np
-from numpy import sin, cos, pi
-import time
+from gym import spaces
 
-from scipy.integrate import odeint
-
-from gym import core, spaces
-from gym.utils import seeding
+from mobileBase.envs.mobileBaseEnv import MobileBaseEnv
 
 
-class MobileBaseAccEnv(core.Env):
-
-    metadata = {"render.modes": ["human", "rgb_array"], "video.frames_per_second": 15}
-
-    BASE_HEIGHT = 1.0 # [m]
-    LINK_MASS_BASE = 500.0  #: [kg] mass of link 1
-
-    MAX_VEL_BASE = 1
-    MAX_POS_BASE = 5.0
-    MAX_ACC_BASE = 1.0
-
-
-    def __init__(self, render=False, dt=0.01):
-        self.viewer = None
-        high = np.array(
-            [
-                self.MAX_POS_BASE,
-                self.MAX_VEL_BASE
-            ],
-            dtype=np.float32,
-        )
-        low = -high
-        self.observation_space = spaces.Box(low=low, high=high, dtype=np.float64)
+class MobileBaseAccEnv(MobileBaseEnv):
+    def setSpaces(self):
+        o = np.concatenate((self._limUpPos, self._limUpVel))
+        self.observation_space = spaces.Box(low=-o, high=o, dtype=np.float64)
         self.action_space = spaces.Box(
-            low=-np.array(self.MAX_ACC_BASE), high=np.array(self.MAX_ACC_BASE), dtype=np.float64
+            low=-self._limUpAcc, high=self._limUpAcc, dtype=np.float64
         )
-        self.state = None
-        self._dt = dt
-        self.seed()
-        self._render = render
-
-    def dt(self):
-        return self._dt
-
-    def seed(self, seed=None):
-        self.np_random, seed = seeding.np_random(seed)
-        return [seed]
-
-    def reset(self):
-        self.state = np.zeros(shape=(2))
-        return self._get_ob()
-
-    def step(self, a):
-        s = self.state
-        self.action = a
-        ns = self.integrate()
-        self.state = np.append(ns, a)
-        terminal = self._terminal()
-        reward = -1.0 if not terminal else 0.0
-        if self._render:
-            self.render()
-        return (self._get_ob(), reward, terminal, {})
-
-    def _get_ob(self):
-        return self.state
-
-    def _terminal(self):
-        if self.state[0] > self.MAX_POS_BASE or self.state[0] < -self.MAX_POS_BASE:
-            return True
-        return False
 
     def continuous_dynamics(self, x, t):
-        u = self.action[0]
-        vel = x[1]
-        acc = np.array([vel, u])
-        return acc
+        vel = x[self._n : self._n * 2]
+        xdot = np.concatenate((vel, self.action))
+        return xdot
 
-    def integrate(self):
-        x0 = self.state[0:2]
-        t = np.arange(0, 2 * self._dt, self._dt)
-        ynext = odeint(self.continuous_dynamics, x0, t)
-        return ynext[1]
-
-    def render(self, mode="human"):
-        from gym.envs.classic_control import rendering
-
-        s = self.state
-
-        if self.viewer is None:
-            self.viewer = rendering.Viewer(500, 500)
-            bound = self.MAX_POS_BASE + 1.0
-            self.viewer.set_bounds(-bound, bound, -bound, bound)
-
-        if s is None:
-            return None
-
-        p0 = [s[0], 0.5 * self.BASE_HEIGHT]
-        p1 = [p0[0], p0[1] + 0.5 * self.BASE_HEIGHT]
-
-        p = [p0, p1]
-        thetas = [0.0, 0.0]
-        tf0 = rendering.Transform(rotation=thetas[0], translation=p0)
-        tf1 = rendering.Transform(rotation=thetas[1], translation=p1)
-        tf = [tf0, tf1]
-
-        self.viewer.draw_line((-5.5, 0), (5.5, 0))
-
-        l, r, t, b = -0.5, 0.5, 0.5, -0.5
-        link = self.viewer.draw_polygon([(l, b), (l, t), (r, t), (r, b)])
-        link.set_color(0, 0.8, 0.8)
-        link.add_attr(tf[0])
-        time.sleep(self.dt())
-
-        return self.viewer.render(return_rgb_array=mode == "rgb_array")
-
-    def close(self):
-        if self.viewer:
-            self.viewer.close()
-            self.viewer = None
